# Restart at most once when prepopulating repository rule environment