@fmeum 2023-12-05
This also allows downloads of `label{_list}` dependencies to be performed in parallel, which is nice.

The ideal solution would probably be to turn `StarlarkBaseExternalContext#getPathFromLabel` into a function that accepts a list of `Label`s and uses `env.getValuesAndExceptions` throughout for "batched" Skyframe requests, but that would require a larger refactoring. This looks like a much simpler fix that should give most of the same benefits.