@Silic0nS0ldier 2023-12-05
> since it does not look like a regression

To confirm, this is not a fix for a regression (at least not a recent one since Bazel 6.4.0 also benefits).

----

> The ideal solution would probably be to turn StarlarkBaseExternalContext#getPathFromLabel into a function that accepts a list of Labels and uses env.getValuesAndExceptions throughout for "batched" Skyframe requests, but that would require a larger refactoring.

Would be more efficient and likely require less code. I think it would make sense to batch that work with resolving [`// TODO(wyv): somehow migrate this to the base context too.`](https://github.com/Silic0nS0ldier/bazel/blob/b8e94f9de10975f2b05ebabee1fb597e76fdc6c1/src/main/java/com/google/devtools/build/lib/bazel/repository/starlark/StarlarkRepositoryContext.java#L518).

Looking at `src/main/java/com/google/devtools/build/lib/bazel/repository/starlark/StarlarkRepositoryFunction.java#fetchInternal` there are plenty of places where `null` is returned (indicating a restart is required). There may yet be more restarts that can be eliminated.

----

I see that this has been tagged for Bazel 7.1. Any chance this can be included in the next v6 release?