@aignas 2023-12-23
@meteorcloudy, since 6.5 has been decided to happen, is this PR something that you would be willing to include in the release? Having these performance improvements could be helpful in the context of rules authors needing to support non-latest bazel version and it may allow to not worry about the restarts as much.