@Silic0nS0ldier 2024-01-17
Tested 6.5.0rc1 locally (macOS) and on CI (Linux), no issues.

Change is working as advertised too. Local test case finished about a minute faster.